requirements.txt
Flask
pymongo
Flask-JWT-Extended
redis
python-dotenv